pyproject.toml
[tool.poetry]
name = "song-graph"
version = "0.1.0"
description = "A graph-based song analysis tool"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "song_graph", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
